chore: refactors function locations
